Revert the Removal of Default Trait Bound
For some reason only whe we use the `Default` trait bound on the `State`
type parameter to the `Fetch` trait do we get a ~20% iteration speed
increase according to ecs_bench.

I prefer the cleanlyness of not requiring the Default trait bound as it
does not as well indicate the intent and the fact that all of the
State's for the different Fetch implementations in that file are going
to have State=(), but it makees more sense to go with the performance
when the readability is not that bad I think.